Significantly reduces required descriptor set count from 40000 to 300 for 100 soldier

+++ src/engine/render_system/RenderTypes.hpp
@@ -132,7 +132,7 @@ struct PipelineGroup {
 
 struct CommandRecordState {
     uint32_t imageIndex;
-    uint8_t frameIndex;
+    uint32_t frameIndex;
     bool isValid = false;
     PipelineGroup * pipeline;
     RenderPass * renderPass;
