+++ src/engine/render_system/render_passes/shadow_render_pass_v2/ShadowRenderPassV2.cpp
@@ -25,7 +25,7 @@ namespace MFA
 
     //-------------------------------------------------------------------------------------------------
 
-    RT::DepthImageGroup const & ShadowRenderPassV2::GetDepthCubeMap(uint8_t const frameIndex) const
+    RT::DepthImageGroup const & ShadowRenderPassV2::GetDepthCubeMap(uint32_t const frameIndex) const
     {
         MFA_ASSERT(mDepthCubeMapList.size() > frameIndex);
         return mDepthCubeMapList[frameIndex];
